Move from a JS file per block registration to one file registering all blocks
This will let webpack combine shared libraries, like gridicons
